refactor(p2): extract editor sheet bundle into sibling page (stage 1 of 2)
SearchServiceEditorSheet + SearchServiceOptionsEditor + SearchAbilityTagLine
plus 16 provider-specific Options Composables (Tavily/Exa/Zhipu/SearXNG/
SearchLinkUp/Brave/Serper/SerpApi/Metaso/Ollama/Perplexity/Firecrawl/Jina/
Bocha/RikkaHub/Grok) moved from SettingSearchPage.kt (L647-940 + L986-1571)
into new sibling SettingSearchServiceEditorSheet.kt.

Byte-equal: bodies and signatures unchanged.
Visibility widened: SearchServiceEditorSheet private -> internal (called by
coordinator across files). SearchAbilityTagLine + ZhipuOptions kept their
pre-existing public visibility byte-equal (no tightening). 16 provider-
specific Options Composables stay private in sibling.

Coordinator strip: 1571 -> 670 (-901 with import cleanup).
Imports trimmed: 19 orphans (rememberScrollState, fillMaxHeight,
KeyboardOptions, verticalScroll, BottomSheetDefaults, ModalBottomSheet,
OutlinedTextField, ProvideTextStyle, SegmentedButton, SegmentedButtonDefaults,
SingleChoiceSegmentedButtonRow, rememberModalBottomSheetState, KeyboardType,
Select, Tag, TagType, SearchService, primaryConstructor, TextButton).
All confirmed this-refactor orphans (>=1 body ref in parent baseline
53604a03, 0 in trimmed coordinator). TextButton had 1 grep hit at L571 but
it's inside a comment only — no real body ref, safe to drop.

CommonOptions left in coordinator for Stage 1 (interleaved between provider
options at parent L942-984); will move to list-cards sibling in Stage 2.

Junction collapse: parent had separator blanks on both sides of the
extracted range (L646 + L941); collapsed to single blank to match P2 stage 1
convention.

External caller note: SearchAbilityTagLine is called from
ui/components/ai/SearchPicker.kt:280 — same package, import line unchanged.

Baseline 6/463 unit tests preserved.
