+++ app/src/main/java/me/rerere/rikkahub/ui/pages/setting/SettingSearchPage.kt
@@ -5,12 +5,10 @@ import me.rerere.hugeicons.stroke.DragDropHorizontal
 import me.rerere.hugeicons.stroke.Add01
 import me.rerere.hugeicons.stroke.Edit01
 import me.rerere.hugeicons.stroke.Delete01
-import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.imePadding
@@ -19,9 +17,6 @@ import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
 import androidx.compose.foundation.lazy.rememberLazyListState
-import androidx.compose.foundation.text.KeyboardOptions
-import androidx.compose.foundation.verticalScroll
-import androidx.compose.material3.BottomSheetDefaults
 import androidx.compose.material3.Card
 import androidx.compose.material3.CardDefaults
 import androidx.compose.material3.Surface
@@ -29,18 +24,10 @@ import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.TopAppBar
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.ModalBottomSheet
-import androidx.compose.material3.OutlinedTextField
-import androidx.compose.material3.ProvideTextStyle
 import androidx.compose.material3.Scaffold
-import androidx.compose.material3.SegmentedButton
-import androidx.compose.material3.SegmentedButtonDefaults
-import androidx.compose.material3.SingleChoiceSegmentedButtonRow
 import androidx.compose.material3.Switch
 import androidx.compose.material3.Text
-import androidx.compose.material3.TextButton
 import androidx.compose.material3.TopAppBarDefaults
-import androidx.compose.material3.rememberModalBottomSheetState
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
@@ -55,7 +42,6 @@ import androidx.compose.ui.hapticfeedback.HapticFeedbackType
 import androidx.compose.ui.input.nestedscroll.nestedScroll
 import androidx.compose.ui.platform.LocalHapticFeedback
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
@@ -66,18 +52,13 @@ import me.rerere.rikkahub.ui.components.nav.BackButton
 import me.rerere.rikkahub.ui.components.ui.AutoAIIcon
 import me.rerere.rikkahub.ui.components.ui.FormItem
 import me.rerere.rikkahub.ui.components.ui.OutlinedNumberInput
-import me.rerere.rikkahub.ui.components.ui.Select
-import me.rerere.rikkahub.ui.components.ui.Tag
-import me.rerere.rikkahub.ui.components.ui.TagType
 import me.rerere.rikkahub.ui.theme.CustomColors
 import me.rerere.rikkahub.utils.plus
 import me.rerere.search.SearchCommonOptions
-import me.rerere.search.SearchService
 import me.rerere.search.SearchServiceOptions
 import org.koin.androidx.compose.koinViewModel
 import sh.calvin.reorderable.ReorderableItem
 import sh.calvin.reorderable.rememberReorderableLazyListState
-import kotlin.reflect.full.primaryConstructor
 import kotlin.uuid.Uuid
 
 @Composable
@@ -645,301 +626,6 @@ private fun SearchProviderCard(
 }
 
 @Composable
-private fun SearchServiceEditorSheet(
-    title: String,
-    confirmText: String,
-    initialService: SearchServiceOptions,
-    onDismiss: () -> Unit,
-    onConfirm: (SearchServiceOptions) -> Unit,
-) {
-    val bottomSheetState = rememberModalBottomSheetState(skipPartiallyExpanded = true)
-    var options by remember(initialService) { mutableStateOf(initialService) }
-
-    ModalBottomSheet(
-        onDismissRequest = onDismiss,
-        sheetState = bottomSheetState,
-        dragHandle = {
-            BottomSheetDefaults.DragHandle()
-        }
-    ) {
-        Column(
-            modifier = Modifier
-                .fillMaxWidth()
-                .fillMaxHeight(0.86f)
-                .padding(16.dp),
-            verticalArrangement = Arrangement.spacedBy(16.dp),
-        ) {
-            Text(
-                text = title,
-                style = MaterialTheme.typography.headlineSmall,
-            )
-
-            Column(
-                modifier = Modifier
-                    .weight(1f)
-                    .verticalScroll(rememberScrollState()),
-                verticalArrangement = Arrangement.spacedBy(12.dp),
-            ) {
-                Select(
-                    options = SearchServiceOptions.TYPES.keys.toList(),
-                    selectedOption = options::class,
-                    optionToString = { SearchServiceOptions.TYPES[it] ?: "[Unknown]" },
-                    onOptionSelected = {
-                        options = it.primaryConstructor!!.callBy(mapOf())
-                    },
-                    optionLeading = {
-                        AutoAIIcon(
-                            name = SearchServiceOptions.TYPES[it] ?: it.simpleName ?: "unknown",
-                            modifier = Modifier.size(24.dp)
-                        )
-                    },
-                    leading = {
-                        AutoAIIcon(
-                            name = SearchServiceOptions.TYPES[options::class] ?: "unknown",
-                            modifier = Modifier.size(24.dp)
-                        )
-                    },
-                    modifier = Modifier.fillMaxWidth(),
-                )
-
-                SearchAbilityTagLine(options = options)
-
-                SearchServiceOptionsEditor(
-                    options = options,
-                    onUpdateOptions = {
-                        options = it
-                    },
-                )
-
-                ProvideTextStyle(MaterialTheme.typography.labelMedium) {
-                    SearchService.getService(options).Description()
-                }
-            }
-
-            Row(
-                modifier = Modifier.fillMaxWidth(),
-                horizontalArrangement = Arrangement.spacedBy(8.dp),
-            ) {
-                TextButton(
-                    onClick = onDismiss,
-                    modifier = Modifier.weight(1f),
-                ) {
-                    Text(stringResource(R.string.cancel))
-                }
-
-                TextButton(
-                    onClick = {
-                        onConfirm(options)
-                    },
-                    modifier = Modifier.weight(1f),
-                ) {
-                    Text(confirmText)
-                }
-            }
-        }
-    }
-}
-
-@Composable
-private fun SearchServiceOptionsEditor(
-    options: SearchServiceOptions,
-    onUpdateOptions: (SearchServiceOptions) -> Unit,
-) {
-    when (options) {
-        is SearchServiceOptions.TavilyOptions -> {
-            TavilyOptions(options) { onUpdateOptions(it) }
-        }
-
-        is SearchServiceOptions.ExaOptions -> {
-            ExaOptions(options) { onUpdateOptions(it) }
-        }
-
-        is SearchServiceOptions.ZhipuOptions -> {
-            ZhipuOptions(options) { onUpdateOptions(it) }
-        }
-
-        is SearchServiceOptions.SearXNGOptions -> {
-            SearXNGOptions(options) { onUpdateOptions(it) }
-        }
-
-        is SearchServiceOptions.LinkUpOptions -> {
-            SearchLinkUpOptions(options) { onUpdateOptions(it) }
-        }
-
-        is SearchServiceOptions.BraveOptions -> {
-            BraveOptions(options) { onUpdateOptions(it) }
-        }
-
-        is SearchServiceOptions.SerperOptions -> {
-            SerperOptions(options) { onUpdateOptions(it) }
-        }
-
-        is SearchServiceOptions.SerpApiOptions -> {
-            SerpApiOptions(options) { onUpdateOptions(it) }
-        }
-
-        is SearchServiceOptions.MetasoOptions -> {
-            MetasoOptions(options) { onUpdateOptions(it) }
-        }
-
-        is SearchServiceOptions.OllamaOptions -> {
-            OllamaOptions(options) { onUpdateOptions(it) }
-        }
-
-        is SearchServiceOptions.PerplexityOptions -> {
-            PerplexityOptions(options) { onUpdateOptions(it) }
-        }
-
-        is SearchServiceOptions.BingLocalOptions -> Unit
-
-        is SearchServiceOptions.FirecrawlOptions -> {
-            FirecrawlOptions(options) { onUpdateOptions(it) }
-        }
-
-        is SearchServiceOptions.JinaOptions -> {
-            JinaOptions(options) { onUpdateOptions(it) }
-        }
-
-        is SearchServiceOptions.BochaOptions -> {
-            BochaOptions(options) { onUpdateOptions(it) }
-        }
-
-        is SearchServiceOptions.RikkaHubOptions -> {
-            RikkaHubOptions(options) { onUpdateOptions(it) }
-        }
-
-        is SearchServiceOptions.GrokOptions -> {
-            GrokOptions(options) { onUpdateOptions(it) }
-        }
-    }
-}
-
-@Composable
-fun SearchAbilityTagLine(
-    modifier: Modifier = Modifier,
-    options: SearchServiceOptions
-) {
-    Row(
-        horizontalArrangement = Arrangement.spacedBy(8.dp),
-        verticalAlignment = Alignment.CenterVertically,
-        modifier = modifier
-    ) {
-        Tag(
-            type = TagType.DEFAULT,
-        ) {
-            Text(stringResource(R.string.search_ability_search))
-        }
-        if (SearchService.getService(options).scrapingParameters != null) {
-            Tag(
-                type = TagType.DEFAULT,
-            ) {
-                Text(stringResource(R.string.search_ability_scrape))
-            }
-        }
-    }
-}
-
-@Composable
-private fun TavilyOptions(
-    options: SearchServiceOptions.TavilyOptions,
-    onUpdateOptions: (SearchServiceOptions.TavilyOptions) -> Unit
-) {
-    FormItem(
-        label = {
-            Text("API Key")
-        }
-    ) {
-        OutlinedTextField(
-            value = options.apiKey,
-            onValueChange = {
-                onUpdateOptions(
-                    options.copy(
-                        apiKey = it
-                    )
-                )
-            },
-            modifier = Modifier.fillMaxWidth()
-        )
-    }
-
-    FormItem(
-        label = {
-            Text("Depth")
-        }
-    ) {
-        val depthOptions = listOf("basic", "advanced")
-        SingleChoiceSegmentedButtonRow(
-            modifier = Modifier.fillMaxWidth()
-        ) {
-            depthOptions.forEachIndexed { index, depth ->
-                SegmentedButton(
-                    shape = SegmentedButtonDefaults.itemShape(index = index, count = depthOptions.size),
-                    onClick = {
-                        onUpdateOptions(
-                            options.copy(
-                                depth = depth
-                            )
-                        )
-                    },
-                    selected = options.depth == depth
-                ) {
-                    Text(depth.replaceFirstChar { it.uppercase() })
-                }
-            }
-        }
-    }
-}
-
-@Composable
-private fun ExaOptions(
-    options: SearchServiceOptions.ExaOptions,
-    onUpdateOptions: (SearchServiceOptions.ExaOptions) -> Unit
-) {
-    FormItem(
-        label = {
-            Text("API Key")
-        }
-    ) {
-        OutlinedTextField(
-            value = options.apiKey,
-            onValueChange = {
-                onUpdateOptions(
-                    options.copy(
-                        apiKey = it
-                    )
-                )
-            },
-            modifier = Modifier.fillMaxWidth()
-        )
-    }
-}
-
-
-@Composable
-fun ZhipuOptions(
-    options: SearchServiceOptions.ZhipuOptions,
-    onUpdateOptions: (SearchServiceOptions.ZhipuOptions) -> Unit
-) {
-    FormItem(
-        label = {
-            Text("API Key")
-        }
-    ) {
-        OutlinedTextField(
-            value = options.apiKey,
-            onValueChange = {
-                onUpdateOptions(
-                    options.copy(
-                        apiKey = it
-                    )
-                )
-            },
-            modifier = Modifier.fillMaxWidth()
-        )
-    }
-}
-
-@Composable
 private fun CommonOptions(
     settings: Settings,
     onUpdate: (SearchCommonOptions) -> Unit
@@ -980,592 +666,5 @@ private fun CommonOptions(
                 )
             }
         }
-    }
-}
-
-@Composable
-private fun SearXNGOptions(
-    options: SearchServiceOptions.SearXNGOptions,
-    onUpdateOptions: (SearchServiceOptions.SearXNGOptions) -> Unit
-) {
-    FormItem(
-        label = {
-            Text("API URL")
-        }
-    ) {
-        OutlinedTextField(
-            value = options.url,
-            onValueChange = {
-                onUpdateOptions(
-                    options.copy(
-                        url = it
-                    )
-                )
-            },
-            modifier = Modifier.fillMaxWidth()
-        )
-    }
-
-    FormItem(
-        label = {
-            Text("Engines")
-        }
-    ) {
-        OutlinedTextField(
-            value = options.engines,
-            onValueChange = {
-                onUpdateOptions(
-                    options.copy(
-                        engines = it
-                    )
-                )
-            },
-            modifier = Modifier.fillMaxWidth()
-        )
-    }
-
-    FormItem(
-        label = {
-            Text("Language")
-        }
-    ) {
-        OutlinedTextField(
-            value = options.language,
-            onValueChange = {
-                onUpdateOptions(
-                    options.copy(
-                        language = it
-                    )
-                )
-            },
-            modifier = Modifier.fillMaxWidth()
-        )
-    }
-
-    FormItem(
-        label = {
-            Text("Username")
-        }
-    ) {
-        OutlinedTextField(
-            value = options.username,
-            onValueChange = {
-                onUpdateOptions(
-                    options.copy(
-                        username = it
-                    )
-                )
-            },
-            modifier = Modifier.fillMaxWidth()
-        )
-    }
-
-    FormItem(
-        label = {
-            Text("Password")
-        }
-    ) {
-        OutlinedTextField(
-            value = options.password,
-            onValueChange = {
-                onUpdateOptions(
-                    options.copy(
-                        password = it
-                    )
-                )
-            },
-            modifier = Modifier.fillMaxWidth()
-        )
-    }
-}
-
-@Composable
-private fun SearchLinkUpOptions(
-    options: SearchServiceOptions.LinkUpOptions,
-    onUpdateOptions: (SearchServiceOptions.LinkUpOptions) -> Unit
-) {
-    FormItem(
-        label = {
-            Text("API Key")
-        }
-    ) {
-        OutlinedTextField(
-            value = options.apiKey,
-            onValueChange = {
-                onUpdateOptions(
-                    options.copy(
-                        apiKey = it
-                    )
-                )
-            },
-            modifier = Modifier.fillMaxWidth()
-        )
-    }
-
-    FormItem(
-        label = {
-            Text("Depth")
-        }
-    ) {
-        val depthOptions = listOf("standard", "deep")
-        SingleChoiceSegmentedButtonRow(
-            modifier = Modifier.fillMaxWidth()
-        ) {
-            depthOptions.forEachIndexed { index, depth ->
-                SegmentedButton(
-                    shape = SegmentedButtonDefaults.itemShape(index = index, count = depthOptions.size),
-                    onClick = {
-                        onUpdateOptions(
-                            options.copy(
-                                depth = depth
-                            )
-                        )
-                    },
-                    selected = options.depth == depth
-                ) {
-                    Text(depth.replaceFirstChar { it.uppercase() })
-                }
-            }
-        }
-    }
-}
-
-@Composable
-private fun BraveOptions(
-    options: SearchServiceOptions.BraveOptions,
-    onUpdateOptions: (SearchServiceOptions.BraveOptions) -> Unit
-) {
-    FormItem(
-        label = {
-            Text("API Key")
-        }
-    ) {
-        OutlinedTextField(
-            value = options.apiKey,
-            onValueChange = {
-                onUpdateOptions(
-                    options.copy(
-                        apiKey = it
-                    )
-                )
-            },
-            modifier = Modifier.fillMaxWidth()
-        )
-    }
-}
-
-@Composable
-private fun SerperOptions(
-    options: SearchServiceOptions.SerperOptions,
-    onUpdateOptions: (SearchServiceOptions.SerperOptions) -> Unit
-) {
-    FormItem(
-        label = {
-            Text("API Key")
-        }
-    ) {
-        OutlinedTextField(
-            value = options.apiKey,
-            onValueChange = {
-                onUpdateOptions(options.copy(apiKey = it))
-            },
-            modifier = Modifier.fillMaxWidth()
-        )
-    }
-}
-
-@Composable
-private fun SerpApiOptions(
-    options: SearchServiceOptions.SerpApiOptions,
-    onUpdateOptions: (SearchServiceOptions.SerpApiOptions) -> Unit
-) {
-    FormItem(
-        label = {
-            Text("API Key")
-        }
-    ) {
-        OutlinedTextField(
-            value = options.apiKey,
-            onValueChange = {
-                onUpdateOptions(options.copy(apiKey = it))
-            },
-            modifier = Modifier.fillMaxWidth()
-        )
-    }
-}
-
-
-@Composable
-private fun MetasoOptions(
-    options: SearchServiceOptions.MetasoOptions,
-    onUpdateOptions: (SearchServiceOptions.MetasoOptions) -> Unit
-) {
-    FormItem(
-        label = {
-            Text("API Key")
-        }
-    ) {
-        OutlinedTextField(
-            value = options.apiKey,
-            onValueChange = {
-                onUpdateOptions(
-                    options.copy(
-                        apiKey = it
-                    )
-                )
-            },
-            modifier = Modifier.fillMaxWidth()
-        )
-    }
-}
-
-@Composable
-private fun OllamaOptions(
-    options: SearchServiceOptions.OllamaOptions,
-    onUpdateOptions: (SearchServiceOptions.OllamaOptions) -> Unit
-) {
-    FormItem(
-        label = {
-            Text("API Key")
-        }
-    ) {
-        OutlinedTextField(
-            value = options.apiKey,
-            onValueChange = {
-                onUpdateOptions(
-                    options.copy(
-                        apiKey = it
-                    )
-                )
-            },
-            modifier = Modifier.fillMaxWidth()
-        )
-    }
-}
-
-@Composable
-private fun PerplexityOptions(
-    options: SearchServiceOptions.PerplexityOptions,
-    onUpdateOptions: (SearchServiceOptions.PerplexityOptions) -> Unit
-) {
-    FormItem(
-        label = {
-            Text("API Key")
-        }
-    ) {
-        OutlinedTextField(
-            value = options.apiKey,
-            onValueChange = {
-                onUpdateOptions(
-                    options.copy(
-                        apiKey = it
-                    )
-                )
-            },
-            modifier = Modifier.fillMaxWidth()
-        )
-    }
-
-    FormItem(
-        label = {
-            Text("Max Tokens")
-        }
-    ) {
-        OutlinedTextField(
-            value = options.maxTokens?.takeIf { it > 0 }?.toString() ?: "",
-            onValueChange = { value ->
-                onUpdateOptions(
-                    options.copy(
-                        maxTokens = value.toIntOrNull()
-                    )
-                )
-            },
-            modifier = Modifier.fillMaxWidth(),
-            keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Number)
-        )
-    }
-
-    FormItem(
-        label = {
-            Text("Max Tokens / Page")
-        }
-    ) {
-        OutlinedTextField(
-            value = options.maxTokensPerPage?.takeIf { it > 0 }?.toString() ?: "",
-            onValueChange = { value ->
-                onUpdateOptions(
-                    options.copy(
-                        maxTokensPerPage = value.toIntOrNull()
-                    )
-                )
-            },
-            modifier = Modifier.fillMaxWidth(),
-            keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Number)
-        )
-    }
-}
-
-@Composable
-private fun FirecrawlOptions(
-    options: SearchServiceOptions.FirecrawlOptions,
-    onUpdateOptions: (SearchServiceOptions.FirecrawlOptions) -> Unit
-) {
-    FormItem(
-        label = {
-            Text("API Key")
-        }
-    ) {
-        OutlinedTextField(
-            value = options.apiKey,
-            onValueChange = {
-                onUpdateOptions(
-                    options.copy(
-                        apiKey = it
-                    )
-                )
-            },
-            modifier = Modifier.fillMaxWidth()
-        )
-    }
-}
-
-@Composable
-private fun JinaOptions(
-    options: SearchServiceOptions.JinaOptions,
-    onUpdateOptions: (SearchServiceOptions.JinaOptions) -> Unit
-) {
-    FormItem(
-        label = {
-            Text("API Key")
-        }
-    ) {
-        OutlinedTextField(
-            value = options.apiKey,
-            onValueChange = {
-                onUpdateOptions(
-                    options.copy(
-                        apiKey = it
-                    )
-                )
-            },
-            modifier = Modifier.fillMaxWidth()
-        )
-    }
-
-    FormItem(
-        label = {
-            Text("Search URL")
-        }
-    ) {
-        OutlinedTextField(
-            value = options.searchUrl,
-            onValueChange = {
-                onUpdateOptions(
-                    options.copy(
-                        searchUrl = it.trim()
-                    )
-                )
-            },
-            modifier = Modifier.fillMaxWidth(),
-            placeholder = {
-                Text("https://s.jina.ai/")
-            }
-        )
-    }
-
-    FormItem(
-        label = {
-            Text("Scrape URL")
-        }
-    ) {
-        OutlinedTextField(
-            value = options.scrapeUrl,
-            onValueChange = {
-                onUpdateOptions(
-                    options.copy(
-                        scrapeUrl = it.trim()
-                    )
-                )
-            },
-            modifier = Modifier.fillMaxWidth(),
-            placeholder = {
-                Text("https://r.jina.ai/")
-            }
-        )
-    }
-}
-
-@Composable
-private fun BochaOptions(
-    options: SearchServiceOptions.BochaOptions,
-    onUpdateOptions: (SearchServiceOptions.BochaOptions) -> Unit
-) {
-    FormItem(
-        label = {
-            Text("API Key")
-        }
-    ) {
-        OutlinedTextField(
-            value = options.apiKey,
-            onValueChange = {
-                onUpdateOptions(
-                    options.copy(
-                        apiKey = it
-                    )
-                )
-            },
-            modifier = Modifier.fillMaxWidth()
-        )
-    }
-
-    FormItem(
-        label = {
-            Text("Summary")
-        },
-        description = {
-            Text("Enable summary generation")
-        },
-        tail = {
-            Switch(
-                checked = options.summary,
-                onCheckedChange = { checked ->
-                    onUpdateOptions(
-                        options.copy(
-                            summary = checked
-                        )
-                    )
-                }
-            )
-        }
-    )
-}
-
-@Composable
-private fun RikkaHubOptions(
-    options: SearchServiceOptions.RikkaHubOptions,
-    onUpdateOptions: (SearchServiceOptions.RikkaHubOptions) -> Unit
-) {
-    FormItem(
-        label = {
-            Text("API Key")
-        }
-    ) {
-        OutlinedTextField(
-            value = options.apiKey,
-            onValueChange = {
-                onUpdateOptions(
-                    options.copy(
-                        apiKey = it
-                    )
-                )
-            },
-            modifier = Modifier.fillMaxWidth()
-        )
-    }
-
-    FormItem(
-        label = {
-            Text("Depth")
-        }
-    ) {
-        val depthOptions = listOf("standard", "deep")
-        SingleChoiceSegmentedButtonRow(
-            modifier = Modifier.fillMaxWidth()
-        ) {
-            depthOptions.forEachIndexed { index, depth ->
-                SegmentedButton(
-                    shape = SegmentedButtonDefaults.itemShape(index = index, count = depthOptions.size),
-                    onClick = {
-                        onUpdateOptions(
-                            options.copy(
-                                depth = depth
-                            )
-                        )
-                    },
-                    selected = options.depth == depth
-                ) {
-                    Text(depth.replaceFirstChar { it.uppercase() })
-                }
-            }
-        }
-    }
-}
-
-@Composable
-private fun GrokOptions(
-    options: SearchServiceOptions.GrokOptions,
-    onUpdateOptions: (SearchServiceOptions.GrokOptions) -> Unit
-) {
-    FormItem(
-        label = {
-            Text("API Key")
-        }
-    ) {
-        OutlinedTextField(
-            value = options.apiKey,
-            onValueChange = {
-                onUpdateOptions(
-                    options.copy(
-                        apiKey = it
-                    )
-                )
-            },
-            modifier = Modifier.fillMaxWidth()
-        )
-    }
-
-    FormItem(
-        label = {
-            Text("Model")
-        }
-    ) {
-        OutlinedTextField(
-            value = options.model,
-            onValueChange = {
-                onUpdateOptions(
-                    options.copy(
-                        model = it
-                    )
-                )
-            },
-            modifier = Modifier.fillMaxWidth()
-        )
-    }
-
-    FormItem(
-        label = {
-            Text("Custom URL")
-        }
-    ) {
-        OutlinedTextField(
-            value = options.customUrl,
-            onValueChange = {
-                onUpdateOptions(
-                    options.copy(
-                        customUrl = it
-                    )
-                )
-            },
-            modifier = Modifier.fillMaxWidth()
-        )
-    }
-
-    FormItem(
-        label = {
-            Text("System Prompt")
-        }
-    ) {
-        OutlinedTextField(
-            value = options.systemPrompt,
-            onValueChange = {
-                onUpdateOptions(
-                    options.copy(
-                        systemPrompt = it
-                    )
-                )
-            },
-            minLines = 3,
-            modifier = Modifier.fillMaxWidth()
-        )
     }
 }
